improve/fix icue sdk process name change

+++ Project-Aurora/Project-Aurora/Settings/Controls/Wrappers/ControlCorsairIcueWrapper.xaml.cs
@@ -95,7 +95,7 @@ public partial class ControlCorsairIcueWrapper
     
     private void UpdateSdkGame()
     {
-        var sdkGamePid = (int)IcueModule.AuroraIcueServer.Sdk.GamePid;
+        var sdkGamePid = IcueModule.AuroraIcueServer.Sdk.GamePid;
         var gameName = IcueModule.AuroraIcueServer.Sdk.GameProcess;
         
         IcueCurrentSdkApplicationLabel.Content = $"{gameName} ({sdkGamePid})";
